+++ app/components/resume/sections-forms/advanced-section-form/SectionFormItem.vue
@@ -41,25 +41,27 @@ const isSummary = computed(() => props.sectionType === "summary")
       @close-edit="closeEditMode"
     />
   </template>
-  <div v-else class="flex items-center gap-1">
-    <UIcon name="i-lucide-grip-vertical" class="text-muted-foreground handle cursor-move" />
-    <button
-      type="button"
-      class="hover:bg-accented/50 bg-accented/30 divide-accented flex flex-1 items-center divide-x rounded-lg p-1 whitespace-pre duration-150 *:px-2 *:text-sm"
-      @click="isEditing = !isEditing"
-    >
-      <span class="line-clamp-1 text-start font-semibold">{{
-        content.title ||
-        ("subtitle" in content && content.subtitle) ||
-        ("location" in content ? content.location : "") ||
-        " "
-      }}</span>
-    </button>
-    <Visibility
-      v-if="sectionType !== 'summary'"
-      :is-hidden="content.isHidden"
-      :on-toggle="() => handleToggleVisibility()"
-    />
-    <Delete v-if="sectionType !== 'summary'" :on-delete="() => emits('delete')" />
+  <div v-else>
+    <div class="flex items-center gap-1">
+      <UIcon name="i-lucide-grip-vertical" class="text-muted-foreground handle shrink-0 cursor-move" />
+      <button
+        type="button"
+        class="hover:bg-accented/50 bg-accented/30 divide-accented line-clamp-1 flex flex-1 items-center divide-x rounded-lg p-1 whitespace-pre duration-150 *:px-2 *:text-sm"
+        @click="isEditing = !isEditing"
+      >
+        <span class="line-clamp-1 text-start font-semibold">{{
+          content.title ||
+          ("subtitle" in content && content.subtitle) ||
+          ("location" in content ? content.location : "") ||
+          " "
+        }}</span>
+      </button>
+      <Visibility
+        v-if="sectionType !== 'summary'"
+        :is-hidden="content.isHidden"
+        :on-toggle="() => handleToggleVisibility()"
+      />
+      <Delete v-if="sectionType !== 'summary'" :on-delete="() => emits('delete')" />
+    </div>
   </div>
 </template>